Moves files in platform-specific folders.
